+++ packages/google_mobile_ads/lib/src/ad_containers.dart
@@ -730,6 +730,11 @@ class InterstitialAd extends AdWithoutView {
   Future<void> show() {
     return instanceManager.showAdWithoutView(this);
   }
+
+  /// Enable / Disable immersive mode for the Ad.
+  Future<void> setImmersiveMode(bool immersiveModeEnabled) async {
+    return instanceManager.setImmersiveMode(this, immersiveModeEnabled);
+  }
 }
 
 /// A full-screen interstitial ad for use with Ad Manager.
@@ -775,6 +780,11 @@ class AdManagerInterstitialAd extends AdWithoutView {
   /// notified of events that occur when showing the ad.
   Future<void> show() {
     return instanceManager.showAdWithoutView(this);
+  }
+
+  /// Enable / Disable immersive mode for the Ad.
+  Future<void> setImmersiveMode(bool immersiveModeEnabled) async {
+    return instanceManager.setImmersiveMode(this, immersiveModeEnabled);
   }
 }
 
@@ -874,6 +884,11 @@ class RewardedAd extends AdWithoutView {
   Future<void> show({required OnUserEarnedRewardCallback onUserEarnedReward}) {
     onUserEarnedRewardCallback = onUserEarnedReward;
     return instanceManager.showAdWithoutView(this);
+  }
+
+  /// Enable / Disable immersive mode for the Ad.
+  Future<void> setImmersiveMode(bool immersiveModeEnabled) async {
+    return instanceManager.setImmersiveMode(this, immersiveModeEnabled);
   }
 }